Create esp_testing.ino
This piece of code is solely for testing purpose of ESP-01 Bluetooth module connectivity to Thingspeak server. The connection of ESP-01 and microcontroller are pretty easy, no GPIO pins of ESP-01 are used. 
VCC and EN ---> 3.3V of MSP430
GND  ---> GND of MSP430 
RX pin ---> TX pin of MSP430
TX pin ---> RX pin of MSP430